+++ SourceCode/AOG/Forms/Guidance/FormHeadLine.cs
@@ -939,7 +939,7 @@ namespace AgOpenGPS
 
         private void cboxToolWidths_SelectedIndexChanged(object sender, EventArgs e)
         {
-            nudSetDistance.Value = (decimal)((Math.Round((mf.tool.width - mf.tool.overlap) * cboxToolWidths.SelectedIndex, 1)) * mf.m2FtOrM);
+            nudSetDistance.Value = Math.Round((mf.tool.width - mf.tool.overlap) * cboxToolWidths.SelectedIndex, 1) * mf.m2FtOrM;
         }
 
         private void btnALength_Click(object sender, EventArgs e)
